+++ marketplace/clients/facebook/client.py
@@ -6,18 +6,16 @@ from django.conf import settings
 from marketplace.clients.base import RequestClient
 
 WHATSAPP_VERSION = settings.WHATSAPP_VERSION
-ACCESS_TOKEN = settings.WHATSAPP_SYSTEM_USER_ACCESS_TOKEN
 
 
 class FacebookAuthorization:
     BASE_URL = f"https://graph.facebook.com/{WHATSAPP_VERSION}/"
 
-    def __init__(self):
-        self.access_token = ACCESS_TOKEN
+    def __init__(self, access_token):
+        self.access_token = access_token
 
     def _get_headers(self):
-        headers = {"Authorization": f"Bearer {self.access_token}"}
-        return headers
+        return {"Authorization": f"Bearer {self.access_token}"}
 
     @property
     def get_url(self):
@@ -137,14 +135,18 @@ class FacebookClient(FacebookAuthorization, RequestClient):
         url = self.get_url + f"{wa_business_id}/owned_product_catalogs"
         headers = self._get_headers()
         all_catalog_ids = []
+        all_catalogs = []
 
         while url:
             response = self.make_request(url, method="GET", headers=headers).json()
             catalog_data = response.get("data", [])
-            all_catalog_ids.extend([item["id"] for item in catalog_data])
+            for item in catalog_data:
+                all_catalog_ids.append(item["id"])
+                all_catalogs.append(item)
+
             url = response.get("paging", {}).get("next")
 
-        return all_catalog_ids
+        return all_catalog_ids, all_catalogs
 
     def destroy_feed(self, feed_id):
         url = self.get_url + f"{feed_id}"