+++ marketplace/core/types/channels/whatsapp_cloud/catalogs/views/views.py
@@ -27,17 +27,27 @@ from marketplace.wpp_products.serializers import (
 
 class BaseViewSet(viewsets.ModelViewSet):
     fb_service_class = FacebookService
+    flows_service_class = FlowsService
+
     fb_client_class = FacebookClient
+    flows_client_class = FlowsClient
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._fb_service = None
+        self._flows_service = None
+
+    def fb_service(self, app: App):  # pragma: no cover
+        access_token = app.apptype.get_access_token(app)
+        if not self._fb_service:
+            self._fb_service = self.fb_service_class(self.fb_client_class(access_token))
+        return self._fb_service
 
     @property
-    def fb_service(self):  # pragma: no cover
-        if not self._fb_service:
-            self._fb_service = self.fb_service_class(self.fb_client_class())
-        return self._fb_service
+    def flows_service(self):  # pragma: no cover
+        if not self._flows_service:
+            self._flows_service = self.flows_service_class(self.flows_client_class())
+        return self._flows_service
 
 
 class Pagination(PageNumberPagination):
@@ -70,7 +80,8 @@ class CatalogViewSet(BaseViewSet):
 
     def retrieve(self, request, *args, **kwargs):
         catalog = self.get_object()
-        connected_catalog_id = self.fb_service.get_connected_catalog(catalog.app)
+        service = self.fb_service(catalog.app)
+        connected_catalog_id = service.get_connected_catalog(catalog.app)
         serializer = self.serializer_class(
             catalog, context={"connected_catalog_id": connected_catalog_id}
         )
@@ -82,9 +93,8 @@ class CatalogViewSet(BaseViewSet):
         serialized_data = []
 
         if queryset.exists():
-            connected_catalog_id = self.fb_service.get_connected_catalog(
-                queryset.first().app
-            )
+            service = self.fb_service(queryset.first().app)
+            connected_catalog_id = service.get_connected_catalog(queryset.first().app)
             serializer = CatalogListSerializer(
                 page_data, context={"connected_catalog_id": connected_catalog_id}
             )
@@ -94,13 +104,29 @@ class CatalogViewSet(BaseViewSet):
 
     @action(detail=True, methods=["POST"])
     def enable_catalog(self, request, *args, **kwargs):
-        response = self.fb_service.enable_catalog(self.get_object())
-        return Response(response)
+        catalog = self.get_object()
+        service = self.fb_service(catalog.app)
+        success, response = service.enable_catalog(catalog)
+        if not success:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data=response)
+
+        self.flows_service.update_catalog_to_active(
+            catalog.app, catalog.facebook_catalog_id
+        )
+        return Response(status=status.HTTP_200_OK)
 
     @action(detail=True, methods=["POST"])
     def disable_catalog(self, request, *args, **kwargs):
-        response = self.fb_service.disable_catalog(self.get_object())
-        return Response(response)
+        catalog = self.get_object()
+        service = self.fb_service(catalog.app)
+        success, response = service.disable_catalog(catalog)
+        if not success:
+            return Response(status=status.HTTP_400_BAD_REQUEST, data=response)
+
+        self.flows_service.update_catalog_to_inactive(
+            catalog.app, catalog.facebook_catalog_id
+        )
+        return Response(status=status.HTTP_200_OK)
 
 
 class CommerceSettingsViewSet(BaseViewSet):
@@ -109,7 +135,8 @@ class CommerceSettingsViewSet(BaseViewSet):
     @action(detail=False, methods=["GET"])
     def commerce_settings_status(self, request, app_uuid, *args, **kwargs):
         app = get_object_or_404(App, uuid=app_uuid, code="wpp-cloud")
-        response = self.fb_service.wpp_commerce_settings(app)
+        service = self.fb_service(app)
+        response = service.wpp_commerce_settings(app)
         return Response(response)
 
     @action(detail=False, methods=["POST"])
@@ -119,7 +146,8 @@ class CommerceSettingsViewSet(BaseViewSet):
         enable_visibility = serializer.validated_data["enable"]
 
         app = get_object_or_404(App, uuid=app_uuid, code="wpp-cloud")
-        response = self.fb_service.toggle_catalog_visibility(app, enable_visibility)
+        service = self.fb_service(app)
+        response = service.toggle_catalog_visibility(app, enable_visibility)
         return Response(response)
 
     @action(detail=False, methods=["POST"])
@@ -129,31 +157,20 @@ class CommerceSettingsViewSet(BaseViewSet):
         enable_cart = serializer.validated_data["enable"]
 
         app = get_object_or_404(App, uuid=app_uuid, code="wpp-cloud")
-        response = self.fb_service.toggle_cart(app, enable_cart)
+        service = self.fb_service(app)
+        response = service.toggle_cart(app, enable_cart)
         return Response(response)
 
     @action(detail=False, methods=["GET"])
     def get_active_catalog(self, request, app_uuid, *args, **kwargs):
         app = get_object_or_404(App, uuid=app_uuid, code="wpp-cloud")
-        response = self.fb_service.get_connected_catalog(app)
+        service = self.fb_service(app)
+        response = service.get_connected_catalog(app)
         return Response(response)
 
 
 class TresholdViewset(BaseViewSet):
     serializer_class = TresholdSerializer
-
-    flows_service_class = FlowsService
-    flows_client_class = FlowsClient
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._flows_service = None
-
-    @property
-    def flows_service(self):  # pragma: no cover
-        if not self._flows_service:
-            self._flows_service = self.flows_service_class(self.flows_client_class())
-        return self._flows_service
 
     @action(detail=True, methods=["POST"])
     def update_treshold(self, request, app_uuid, *args, **kwargs):