+++ marketplace/wpp_templates/views.py
@@ -18,6 +18,7 @@ from rest_framework.pagination import PageNumberPagination
 from django.core.exceptions import ValidationError
 
 from marketplace.applications.models import App
+from marketplace.core.types import APPTYPES
 from marketplace.core.types.channels.whatsapp_base.exceptions import (
     FacebookApiException,
 )
@@ -116,7 +117,7 @@ class TemplateMessageViewSet(viewsets.ModelViewSet):
                     f"This app: {instance.app.uuid} does not have fb_access_token in config"
                 )
         else:
-            access_token = settings.WHATSAPP_SYSTEM_USER_ACCESS_TOKEN
+            access_token = APPTYPES.get("wpp-cloud").get_access_token(instance.app)
             waba_id = instance.app.config.get("wa_waba_id")
 
         if waba_id is None:
@@ -171,7 +172,7 @@ class TemplateMessageViewSet(viewsets.ModelViewSet):
                     f"This app: {template.app.uuid} does not have fb_access_token in config"
                 )
         else:
-            access_token = settings.WHATSAPP_SYSTEM_USER_ACCESS_TOKEN
+            access_token = APPTYPES.get("wpp-cloud").get_access_token(template.app)
 
         message_template_id = request.data.get("message_template_id")
 
@@ -200,7 +201,7 @@ class TemplateMessageViewSet(viewsets.ModelViewSet):
                 or header.get("format") == "VIDEO"
             ):
                 photo_api_request = PhotoAPIRequest(
-                    template.app.config.get("wa_waba_id")
+                    access_token, template.app.config.get("wa_waba_id")
                 )
                 photo = header.get("example")
                 file_type = re.search("(?<=data:)(.*)(?=;base64)", photo).group(0)